Move FAB animation logic into LavaFloatingActionButton.kt
